fix(react): react tag references variable scope
